+++ test/klippy/z_tilt_ng.test
@@ -1,0 +1,28 @@
+# Test case for z_tilt_ng
+CONFIG z_tilt_ng.cfg
+DICTIONARY atmega2560.dict
+
+# Start by homing the printer.
+G28
+G1 Z5 X10 Y10 F6000
+M400
+GET_POSITION
+
+# Run Z_TILT_ADJUST in manual mode
+Z_TILT_ADJUST METHOD=MANUAL
+G1 Z2.909972
+ACCEPT
+G1 Z2.924972
+ACCEPT
+G1 Z2.959972
+ACCEPT
+G1 Z2.924972
+ACCEPT
+
+# Report position
+G1 Z5 X10 Y10 F6000
+M400
+GET_POSITION
+
+# Run again in automatic mode
+Z_TILT_ADJUST
